+++ example/lib/screens/settings_screen.dart
@@ -61,7 +61,6 @@ class _SettingsScreenState extends State<SettingsScreen> {
                 title: Text('Lock app in background'),
                 leading: Icon(Icons.phonelink_lock),
                 switchValue: lockInBackground,
-                switchActiveColor: Colors.green,
                 onToggle: (bool value) {
                   setState(() {
                     lockInBackground = value;
@@ -102,7 +101,26 @@ class _SettingsScreenState extends State<SettingsScreen> {
                 leading: Icon(Icons.collections_bookmark),
               ),
             ],
-          )
+          ),
+          CustomSection(
+            child: Column(
+              children: [
+                Padding(
+                  padding: const EdgeInsets.only(top: 22, bottom: 8),
+                  child: Image.asset(
+                    'assets/settings.png',
+                    height: 50,
+                    width: 50,
+                    color: Color(0xFF777777),
+                  ),
+                ),
+                Text(
+                  'Version: 2.4.0 (287)',
+                  style: TextStyle(color: Color(0xFF777777)),
+                ),
+              ],
+            ),
+          ),
         ],
       ),
     );